pyproject.toml
[tool.poetry]
name = "qna-helper"
version = "0.1.0"
description = "CLI-utilite builded for answering question in csv using davinci-003"
authors = ["emp7yhead <artyomkropp@gmail.com>"]
readme = "README.md"
packages = [{include = "qna_helper"}]

[tool.poetry.scripts]
qna-helper="qna_helper.scripts.qna_helper:main"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.13.3"
alive-progress = "^3.1.5"
python-dotenv = "^1.0.1"
tqdm = "^4.66.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
